Added new sample - test todo app; divided into several projects; udated expected status code method to use array
