--- framework/src/main/java/io/test/automation/cucumberframework/context/DefaultContextConfiguration.java
@@ -1,9 +1,0 @@
-package io.test.automation.cucumberframework.context;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan(basePackages = {"io.test.automation"})
-public class DefaultContextConfiguration {
-}
