+++ framework/src/main/java/io/test/automation/cucumberframework/general/context/DefaultContextConfiguration.java
@@ -1,0 +1,9 @@
+package io.test.automation.cucumberframework.general.context;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan(basePackages = {"io.test.automation"})
+public class DefaultContextConfiguration {
+}
