--- framework/src/main/java/io/test/automation/cucumberframework/models/RegistrationAndLoginResponseDto.java
@@ -1,13 +1,0 @@
-package io.test.automation.cucumberframework.models;
-
-import lombok.Data;
-import lombok.experimental.Accessors;
-
-@Data
-@Accessors(chain = true)
-public class RegistrationAndLoginResponseDto {
-
-    private Long code;
-    private String message;
-    private RegistrationResponseDataDto data;
-}
