--- framework/src/main/java/io/test/automation/cucumberframework/models/RegistrationResponseDataDto.java
@@ -1,19 +1,0 @@
-package io.test.automation.cucumberframework.models;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-import lombok.experimental.Accessors;
-
-@Data
-@Accessors(chain = true)
-public class RegistrationResponseDataDto {
-
-    @JsonProperty("Id")
-    private Long id;
-    @JsonProperty("Name")
-    private String name;
-    @JsonProperty("Email")
-    private String email;
-    @JsonProperty("Token")
-    private String token;
-}
