--- framework/src/main/java/io/test/automation/cucumberframework/pages/MyAccountPage.java
@@ -1,12 +1,0 @@
-package io.test.automation.cucumberframework.pages;
-
-public class MyAccountPage extends AbstractPage {
-
-    public static final String PAGE_HEADER = "//p[text()='My account']";
-    private static final String SUCCESS_ALERT = ".alert-success";
-    private static final String HOME_BTN = "//a[@title='Home']";
-
-    public MyAccountPage() {
-        super();
-    }
-}
