+++ framework/src/main/java/io/test/automation/cucumberframework/sampleshopapp/models/LoginRequestDto.java
@@ -1,0 +1,12 @@
+package io.test.automation.cucumberframework.sampleshopapp.models;
+
+import lombok.Data;
+import lombok.experimental.Accessors;
+
+@Data
+@Accessors(chain = true)
+public class LoginRequestDto {
+
+    private String email;
+    private String password;
+}
