+++ framework/src/main/java/io/test/automation/cucumberframework/sampleshopapp/models/RegistrationResponseDataDto.java
@@ -1,0 +1,19 @@
+package io.test.automation.cucumberframework.sampleshopapp.models;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+import lombok.experimental.Accessors;
+
+@Data
+@Accessors(chain = true)
+public class RegistrationResponseDataDto {
+
+    @JsonProperty("Id")
+    private Long id;
+    @JsonProperty("Name")
+    private String name;
+    @JsonProperty("Email")
+    private String email;
+    @JsonProperty("Token")
+    private String token;
+}
