+++ framework/src/main/java/io/test/automation/cucumberframework/sampleshopapp/models/UserDto.java
@@ -1,0 +1,16 @@
+package io.test.automation.cucumberframework.sampleshopapp.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class UserDto {
+
+    private Long id;
+    private String name;
+    private String email;
+    private String profilepicture;
+    private String location;
+    private String createdat;
+}
